fixed wrap around bug of stepAuxDelay higher than 32ms of xybot

+++ firmwares/xybot/xybot.ino
@@ -73,7 +73,7 @@ void stepperMoveB(int dir)
 /************** calculate movements ******************/
 //#define STEPDELAY_MIN 200 // micro second
 //#define STEPDELAY_MAX 1000
-int stepAuxDelay=0;
+long stepAuxDelay=0; // int not sufficient because only supports delay up to 32ms
 int stepdelay_min=200;
 int stepdelay_max=1000;
 #define ACCELERATION 2 // mm/s^2 don't get inertia exceed motor could handle
@@ -82,7 +82,7 @@ int stepdelay_max=1000;
 
 void doMove()
 {
-  int mDelay=stepdelay_max;
+  long mDelay=stepdelay_max; // int not sufficient because only supports delay up to 32ms
   int speedDiff = -SPEED_STEP;
   int dA,dB,maxD;
   float stepA,stepB,cntA=0,cntB=0;
@@ -201,7 +201,7 @@ void parseCordinate(char * cmd)
       float speed = atof(str+1);
       tarSpd = speed/60; // mm/min -> mm/s
     }else if(str[0]=='A'){
-      stepAuxDelay = atoi(str+1);
+      stepAuxDelay = atol(str+1);
     }
   }
   prepareMove();
@@ -270,7 +270,7 @@ void parseAuxDelay(char * cmd)
 {
   char * tmp;
   strtok_r(cmd, " ", &tmp);
-  stepAuxDelay = atoi(tmp);
+  stepAuxDelay = atol(tmp);
 }
 
 void parseLaserPower(char * cmd)
